moved pkg's into their own domain folders: server, mysql; removed pkg folder
