+++ go.mod
@@ -7,10 +7,8 @@ require (
 	github.com/jinzhu/gorm v1.9.8
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/labstack/echo v3.3.10+incompatible
-	github.com/labstack/gommon v0.2.8 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
+	github.com/labstack/gommon v0.2.9 // indirect
 	github.com/pkg/errors v0.8.1
 	github.com/spf13/cobra v0.0.4
 	github.com/spf13/viper v1.3.2
-	github.com/valyala/fasttemplate v1.0.1 // indirect
 )
