+++ mysql/models.go
@@ -1,4 +1,4 @@
-package db
+package mysql
 
 import (
 	"log"
